Started implementing Simplification. In particular, the algorithms used by SimplificationRule have been implemented.
In addition, some new interfaces have been added to allow usage of more general usage of certain functions, e.g. Summable, Multipliable, and Powerable are interfaces to allow Sum, Product, and Power to be implemented more generally.

Removal of Topology as current implementation is not useful. Addition of some basic functions to demonstrate usage of operation tree. As of now, main program has an example of a simplification rule in action.